fix: resolve all critical and high audit findings across all phases
Security (6 critical/high fixes):
- Prevent transaction hijacking: verify ownership on commit/rollback/execute
- Use cryptographically random transaction IDs instead of sequential counter
- Raise PBKDF2-HMAC-SHA256 iterations from 10k to 600k (OWASP minimum)
- Deny-by-default for unhandled statement types in authorize_statement()
  (was catch-all Allow; now explicitly handles AlterTable, Explain)
- Authenticate get_server_info endpoint (was unauthenticated info disclosure)
- Permissive mode no longer grants superuser role (returns 'default' role)

Storage Engine (5 critical/high fixes):
- Fix TOCTOU race: move maybe_freeze_memtable inside write_mutex in put/delete/write
- FilePager: add sync_data() after write_page for crash durability
- FilePager: add CRC32 checksums on disk pages (9-byte envelope header)
- FilePager: use read_exact() instead of read() to prevent partial reads
- SSTable fsync before WAL deletion to close data loss window
- Fix compaction dedup: sort by (key ASC, sequence DESC) to keep newest entry

io_uring (4 high fixes, Linux-only):
- Hold registered_buffers lock for entire SQE lifecycle (was use-after-free)
- Page-align registered buffers to 4096 for O_DIRECT compatibility
- Replace mem::forget with fd=-1 sentinel in close() to prevent resource leak
- Mark UringBatch::add_read/add_write as unsafe (buffer lifetime not enforced)
- Use target_fd() in read_at/write_at to respect fixed-fd mode

Caching (6 high fixes):
- Fix ResultCache deadlock: consistent lock ordering (cache before table_index)
- Fix table_index memory leak: trim stale entries after LRU eviction
- SQL normalization now preserves string literal case (was lowercasing 'Alice')
- invalidate_where now clears cache (was a no-op)
- CountingBloomFilter: u8 -> u16 counters to prevent false negatives on overflow
- Document evict_expired() lazy strategy

Vector Database (4 high fixes):
- Filter deleted/tombstoned nodes from HNSW search results
- INSERT now populates HNSW vector indexes automatically
- CREATE INDEX now backfills existing rows into new HNSW index
- DROP DATABASE cascades to vector index cleanup

+++ crates/nexus-cache/src/bloom.rs
@@ -212,8 +212,8 @@ impl std::fmt::Debug for BloomFilter {
 /// Instead of single bits, uses counters so items can be removed.
 /// Uses more memory than a standard Bloom filter.
 pub struct CountingBloomFilter {
-    /// Counter array.
-    counters: Vec<u8>,
+    /// Counter array (u16 to avoid overflow after 256+ insertions of the same item).
+    counters: Vec<u16>,
     /// Number of counters.
     num_counters: usize,
     /// Number of hash functions.
@@ -239,7 +239,7 @@ impl CountingBloomFilter {
             .collect();
 
         Self {
-            counters: vec![0u8; num_counters],
+            counters: vec![0u16; num_counters],
             num_counters,
             num_hashes,
             count: 0,
